Make JsonOperation have an instance for VIZ use and separate the TableVizOperation to own class.
