+++ distribution/lib/Standard/Table/0.0.0-dev/src/Row.enso
@@ -99,8 +99,7 @@ type Row
        private: true
        ---
        Converts this row into a JSON representation.
-    to_json_data : Text
-    to_json_data self =
+    to_json_data self -> Text =
         self.java_row.toJsonData
 
 ## ---
