+++ std-bits/table/src/main/java/org/enso/table/data/table/Table.java
@@ -19,6 +19,7 @@ import org.enso.table.aggregations.Aggregator;
 import org.enso.table.data.column.builder.Builder;
 import org.enso.table.data.column.operation.JsonOperation;
 import org.enso.table.data.column.operation.StorageIterators;
+import org.enso.table.data.column.operation.TableVizOperation;
 import org.enso.table.data.column.operation.masks.IndexMapper;
 import org.enso.table.data.column.storage.ColumnBooleanStorage;
 import org.enso.table.data.column.storage.ColumnStorage;
@@ -670,7 +671,7 @@ public final class Table {
 
   public String tableVizJSON(
       List<String> valueTypeDisplay, long allRowsCount, boolean useServerMode) {
-    return JsonOperation.makeTableVizJSON(
+    return TableVizOperation.makeJSON(
         versionId, columns, allRowsCount, useServerMode, valueTypeDisplay, "get_row");
   }
 }
